Move winapi-related implementation to the existing winapi module
